# cmd/api/docs/swagger.yaml
definitions:
...
  entity.LoginResponse:
    properties:
      token:
        type: string
    type: object
  entity.ResponseData:
    properties:
      data: {}
    type: object
  entity.ResponseError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  entity.UserCreateRequest:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
  entity.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  entity.UserResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  entity.UserUpdateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/entity.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/entity.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ResponseError'
      summary: User login
      tags:
      - auth
...
  /management/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ResponseError'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - management users
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: Create user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ResponseError'
      summary: Create a new user
      tags:
      - users
  /users/info:
    get:
      consumes:
      - application/json
      description: Get a user by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ResponseError'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user with the input payload
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/entity.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ResponseError'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"